requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg[binary]==3.2.*
qrcode==7.4.2
Pillow==10.4.0
alembic